Updated program documentation and added accent removal for characters
Tested all the input text files and saved the output (via screenshots). The code is finalized and now has accent removal for any accents that may be encountered when generating art.

Project is 90% complete, just need to complete the 'Discussion' section in the README and record the presentation of the system.